refactor : 클래스가 하나만 있는 controller,domain 패키지 삭제
