+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,5 @@
 package racingcar;
 
-import racingcar.controller.RaceController;
-
 public class Application {
 	public static void main(String[] args) {
 		// TODO 구현 진행
